+++ awb_aradial_connector/api/aradial_gateway/update_user.py
@@ -2,7 +2,6 @@ import json
 import requests
 from odoo import exceptions
 from requests.auth import HTTPBasicAuth
-from .get_user import AradialAPIGatewayGetUser
 
 import logging
 
@@ -12,14 +11,12 @@ class AradialAPIGatewayUpdateUser(object):
     def __init__(
         self,
         url,
-        userbalance_url,
         username,
         password,
         data
     ):
 
         self.url = url
-        self.userbalance_url = userbalance_url
         self.username = username
         self.password = password
 
@@ -33,13 +30,6 @@ class AradialAPIGatewayUpdateUser(object):
     def update_user(self):
         _logger.info('function: update_user')
 
-        # Update User's Product
-        offer_data = {
-            'Offer': self.data['Offer'],
-            'Status': '0',
-            'CustomInfo1':self.data['CustomInfo1']
-        }
-
         try:
             res = requests.put(
                 url=self.url+'/'+self.data['UserID'],
@@ -51,40 +41,9 @@ class AradialAPIGatewayUpdateUser(object):
             raise exceptions.ValidationError(e)
 
         if res.status_code != 204:
-            update_offer_state = False
+            update_state = False
             _logger.error('!!! Error Updating Offer to '+self.data['Offer']+' for Subscriber '+self.data['UserID'])
         else:
-            update_offer_state = True
+            update_state = True
 
-
-        gateway = AradialAPIGatewayGetUser(self.url, self.username, self.password)
-        user = gateway.get_user(self.data['UserID'])
-
-        if user['Offer'] == self.data['Offer']:
-            # Update User's TimeBank
-            if self.data['Timebank'] > 0:
-                timebank_data = {
-                    'TimeBank': self.data['Timebank']
-                }
-
-                try:
-                    res = requests.post(
-                        url=self.userbalance_url+'/'+self.data['UserID'],
-                        headers=self.headers,
-                        data=json.dumps(timebank_data),
-                        auth=HTTPBasicAuth(self.username, self.password)
-                    )
-                except requests.exceptions.MissingSchema as e:
-                    raise exceptions.ValidationError(e)
-
-                if res.status_code != 201:
-                    update_timebank_state = False
-                    _logger.error('!!! Error Adding to TimeBank for Subscriber '+self.data['UserID'])
-                else:
-                    update_timebank_state = True
-            
-            return True if update_offer_state and update_timebank_state else False
-
-        return True if update_offer_state else False
-
-
+        return update_state