+++ awb_aradial_connector/models/awb_aradial_connector.py
@@ -66,17 +66,15 @@ class AWBAradialConnector(models.Model):
         self, 
         data
     ):
-        _logger.info("Update User's Product and TimeBank")
+        _logger.info("Update User")
         
         params = self.env['ir.config_parameter'].sudo()
         self.aradial_url = params.get_param('aradial_url')
-        self.aradial_userbalance_url = params.get_param('aradial_userbalance_url')
         self.aradial_username = params.get_param('aradial_username')
         self.aradial_password = params.get_param('aradial_password')
 
         user = AradialAPIGatewayUpdateUser(
             url=self.aradial_url,
-            userbalance_url = self.aradial_userbalance_url,
             username=self.aradial_username,
             password=self.aradial_password,
             data=data