+++ awb_subscriber_bill_automation/api/client/sf_update_account.py
@@ -1,6 +1,8 @@
 from odoo import api, fields, models, _
 from odoo.exceptions import UserError
 from datetime import datetime
+from pytz import timezone
+
 
 import logging
 
@@ -11,6 +13,14 @@ class SalesForceConnector(models.Model):
 
     def update_account(self, record=None, update_type=None, main_plan=None):
         _logger.info('function: SalesForceConnector => update_account')
+
+        # timezone = self.env.user.tz or pytz.utc
+        # now = datetime.datetime.now(pytz.timezone(timezone))
+        
+        # Current time in UTC
+        now_utc = datetime.now(timezone('UTC'))
+        # Convert to Asia/Manila time zone
+        now = now_utc.astimezone(timezone('Asia/Manila'))
         
         if record:
             self.data = {}
@@ -33,14 +43,16 @@ class SalesForceConnector(models.Model):
                 _logger.info('Entered update_type equal to 6')
                 self.data = {
                     'SFID': record.opportunity_id.salesforce_id,
-                    'SMS_Activation_Date_Time': datetime.now().strftime("%m/%d/%Y %I:%M%p"),
+                    'BillCustomerID': record.customer_number,
+                    'SMS_Activation_Date_Time': now.strftime("%m/%d/%Y %I:%M%p"),
+                    # 'SMS_Activation_Date_Time': datetime.now().strftime("%m/%d/%Y %I:%M%p"),
                     'SubscriptionCode': record.code,
                     'AccountStatus': "Active",
                     'ProductCode': main_plan.default_code.upper(),
                     'UpdateType': update_type
                 }
             elif update_type == 7:
-                _logger.info('Entered update_type equal to 6')
+                _logger.info('Entered update_type equal to 7')
                 self.data = {
                     'SFID': record.opportunity_id.salesforce_id,
                     'SubscriptionCode': record.code,
@@ -58,6 +70,4 @@ class SalesForceConnector(models.Model):
   
         _logger.info(f'SF Update Account Status: {update_account_stat}')
         return update_account_stat
-
-