+++ awb_subscriber_bill_automation/models/subscription.py
@@ -9,6 +9,8 @@ from odoo import api, fields, models, _
 from odoo.exceptions import UserError
 from dateutil.relativedelta import relativedelta
 from datetime import datetime
+from  pytz import timezone
+
 import json
 
 import logging
@@ -44,7 +46,7 @@ class SaleSubscription(models.Model):
 
     # New Fields
 
-    plan_type = fields.Many2one('product.plan.type', compute='_compute_plan_type', store=True)
+    plan_type = fields.Many2one('product.plan.type', compute='_compute_plan_type')
 
     # Business Logic
 
@@ -67,7 +69,7 @@ class SaleSubscription(models.Model):
         # vals['atm_ref_sequence'] = self.env['ir.sequence'].next_by_code('subscription.atm.reference.seq.code')
         res = super(SaleSubscription, self).create(vals)
 
-        _logger.info('function: create')
+        _logger.info('SMS::function: create')
         self.record = res
         sms_flag = True
         plan_type = ''
@@ -76,14 +78,14 @@ class SaleSubscription(models.Model):
 
         try:
             main_plan = self._get_mainplan(self.record)        
-            _logger.debug(f'Main_Plan: {main_plan}')
+            _logger.debug(f'SMS::Main_Plan: {main_plan}')
 
             plan_type = main_plan.sf_plan_type.name
             if plan_type == 'Postpaid':
                 sms_flag = False
 
         except:
-            _logger.warning("Main Plan not found")
+            _logger.warning("SMS:: Main Plan not found")
             sms_flag = False
 
         if sms_flag and plan_type == 'Prepaid':
@@ -103,12 +105,12 @@ class SaleSubscription(models.Model):
                     subtype = "disconnection-temporary"
                     self.env['sale.subscription']._change_status_subtype(last_subscription, subtype, is_closed_subs, ctp)
                 except:
-                    _logger.error(f'!!! Failed Temporary Discon - Subscription code {self.record.code}')
+                    _logger.error(f'SMS:: !!! Failed Temporary Discon - Subscription code {self.record.code}')
 
-            # self.env['sale.subscription'].provision_and_activation(self.record, main_plan, last_subscription, ctp)
+            self.env['sale.subscription'].provision_and_activation(self.record, main_plan, last_subscription, ctp)
 
             # Helper to update Odoo Opportunity
-            # self._update_account(main_plan, self.record, sf_update_type, max_fail_retries)            
+            self._update_account(main_plan, self.record, sf_update_type, max_fail_retries)            
 
         if not ctp:
             self.env['sale.subscription'].generate_atmref(self.record, max_fail_retries)
@@ -117,7 +119,7 @@ class SaleSubscription(models.Model):
 
     # TODO For Clean up, refer to compute plan type
     def _get_mainplan(self, record):
-        _logger.info('function: get_mainplan')
+        _logger.info('SMS:: function: get_mainplan')
 
         main_plan = ''
 
@@ -132,11 +134,12 @@ class SaleSubscription(models.Model):
 
 
     def _get_last_subscription(self, record, plan_type):
-        _logger.info('function: _get_last_subscription')
+        _logger.info('SMS:: function: _get_last_subscription')
         customer_id = record.customer_number
 
-        _logger.debug(f'Subscription Code: {record.code}')
-        _logger.debug(f'Customer Number: {customer_id}')
+        _logger.debug(f'SMS:: Subscription Code: {record.code}')
+        
+        _logger.debug(f'SMS:: Customer Number: {customer_id}')
 
         last_subscription = False
         subscriptions = self.env['sale.subscription'].search([('customer_number','=', customer_id),('plan_type','=', plan_type)], order='id desc', limit=2)
@@ -147,18 +150,20 @@ class SaleSubscription(models.Model):
 
         return last_subscription
 
+
     def _update_account(self, main_plan, rec, sf_update_type, max_retries):
+        _logger.debug(f'SMS:: _update_account')
         try:
             self.env['sale.subscription'].update_account(rec, sf_update_type, main_plan)
         except:
             if max_retries > 1:
                 self._update_account(main_plan, rec, sf_update_type, max_retries-1)
             else:
-                _logger.error(f'!!! Failed SF Update Account Status - Subscription code {rec.code}')
+                _logger.error(f'SMS:: !!! Failed SF Update Account Status - Subscription code {rec.code}')
                 raise Exception(f'!!! Failed SF Update Account Status - Subscription code {rec.code}')
     
     def _update_new_subscription(self, record, last_subscription):
-        _logger.info('function: _update_new_subscription')
+        _logger.info('SMS:: function: _update_new_subscription')
 
         self.record = record
         self.record['opportunity_id'] = last_subscription.opportunity_id.id
@@ -166,7 +171,7 @@ class SaleSubscription(models.Model):
         
         self.env.cr.commit()
 
-        _logger.debug(f'New Subscription: {self.record}')
+        _logger.debug(f'SMS:: New Subscription: {self.record}')
         return self.record
 
 
@@ -182,6 +187,7 @@ class SaleSubscription(models.Model):
 
     @api.depends("atm_ref_sequence")
     def _compute_atm_reference_number(self):
+        _logger.debug(f'SMS:: _compute_atm_reference_number')
         for rec in self:
             rec.atm_ref = ''
             if rec.atm_ref_sequence:
@@ -494,8 +500,23 @@ class SaleSubscription(models.Model):
 
 
     def _get_datetime_now(self):
-        for rec in self:
+        _logger.info(f'SMS:: function: _get_datetime_now')    
+        try:
+            # timezone = self.env.user.tz or pytz.utc
+            # now = datetime.now(pytz.timezone(timezone))
+
+            # Current time in UTC
+            now_utc = datetime.now(timezone('UTC'))
+            # Convert to Asia/Manila time zone
+            now = now_utc.astimezone(timezone('Asia/Manila'))
+
+            for rec in self:
+                rec.datetime_now = now.strftime("%m/%d/%Y %I:%M %p")
+                _logger.debug(f'SMS::rec.datetime_now {rec.datetime_now}')
+        except:
+            _logger.error(f'SMS:: Error encountered in getting date and time..') 
             rec.datetime_now = datetime.now().strftime("%m/%d/%Y %I:%M %p")
+    
 
 
 class SaleSubscriptionLine(models.Model):