+++ awb_aradial_connector/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [13.0.1.0.2] - 2021-09-22
+- [FIX] Removed update to Timebank
+
 ## [13.0.1.0.1] - 2021-09-21
 - [IMP] Added code to update user in Aradial