+++ awb_subscriber_bill_automation/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.5.7] - 2021-09-22
+- [FIX] Datetime(timezone) passed from Odoo to SF in Update Account
+- [FIX] Datetime(timezone) in Expiry Notification
+- [ADD] Start and End Date Subscription Prepaid
+- [MISC] Updated implementation of _get_last_subscrption
+
+## [13.0.1.5.6] - 2021-09-21
+- [FIX] Temporary discon due to Expiry - Stage InProgress to Closed
+- [ADD] Activation Portal
+- [FIX] Update New Subscription with Opportunity ID from Last Subscription
+- [FIX] Subscription Stage changed from In Progress to Closed
+
 ## [13.0.1.5.5] - 2021-09-21
 - [FIX] Update New Subscription with Opportunity ID from Last Subscription
 - [IMP] ATM Ref Generation : Inherit from the last Subscription ATM Ref
@@ -14,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ADD] Added plan type in Subscription record
 - [FIX] Temporary discon due to Expiry - Stage InProgress to Closed
 - [ADD] Activation Portal
-
 
 ## [13.0.1.5.4] - 2021-09-20
 - [FIX] Populate PrepaidIndicator flag in the Aradial Create User payload